+++ booking-app/app/api/bookings/route.ts
@@ -2,6 +2,7 @@ import { toFirebaseTimestampFromString } from "@/components/src/client/utils/ser
 import {
   firstApproverEmails,
   serverApproveInstantBooking,
+  serverBookingContents,
   serverSendBookingDetailEmail,
   serverUpdateDataByCalendarEventId,
 } from "@/components/src/server/admin";
@@ -19,6 +20,7 @@ import {
 } from "@/components/src/types";
 import {
   logServerBookingChange,
+  serverFetchAllDataFromCollection,
   serverGetNextSequentialId,
   serverSaveDataToFirestore,
   serverGetDocumentById,
@@ -243,6 +245,102 @@ async function createBookingCalendarEvent(
   });
   return event.id;
 }
+
+const SERVICE_APPROVER_CONFIG = {
+  setup: {
+    flagField: "isSetup",
+    subjectStatus: "SETUP REQUESTED",
+    displayName: "setup",
+  },
+  equipment: {
+    flagField: "isEquipment",
+    subjectStatus: "EQUIPMENT REQUESTED",
+    displayName: "equipment",
+  },
+  staff: {
+    flagField: "isStaffing",
+    subjectStatus: "STAFFING REQUESTED",
+    displayName: "staffing",
+  },
+  catering: {
+    flagField: "isCatering",
+    subjectStatus: "CATERING REQUESTED",
+    displayName: "catering",
+  },
+  cleaning: {
+    flagField: "isCleaning",
+    subjectStatus: "CLEANUP REQUESTED",
+    displayName: "cleanup",
+  },
+  security: {
+    flagField: "isSecurity",
+    subjectStatus: "SECURITY REQUESTED",
+    displayName: "security",
+  },
+} as const;
+
+const notifyServiceApproversForRequestedServices = async (
+  calendarEventId: string,
+  requesterEmail: string,
+  servicesRequested: ReturnType<typeof getMediaCommonsServices>,
+  tenant: string,
+) => {
+  if (!Object.values(servicesRequested).some(Boolean)) {
+    return;
+  }
+
+  const usersRights = await serverFetchAllDataFromCollection<any>(
+    TableNames.USERS_RIGHTS,
+    [],
+    tenant,
+  );
+  const bookingContents = await serverBookingContents(calendarEventId, tenant);
+  const bookingContentsAsStrings = Object.fromEntries(
+    Object.entries(bookingContents).map(([key, value]) => [
+      key,
+      value instanceof Timestamp ? value.toDate().toISOString() : String(value ?? ""),
+    ]),
+  );
+  const emailConfig = await getTenantEmailConfig(tenant);
+
+  const emailJobs = Object.entries(SERVICE_APPROVER_CONFIG).flatMap(
+    ([serviceKey, config]) => {
+      if (!servicesRequested[serviceKey as keyof typeof servicesRequested]) {
+        return [];
+      }
+
+      const recipients = Array.from(
+        new Set(
+          usersRights
+            .filter((record) => record[config.flagField] === true)
+            .map((record) => record.email)
+            .filter(Boolean),
+        ),
+      );
+
+      return recipients.map((recipient) =>
+        sendHTMLEmail({
+          templateName: "booking_detail",
+          contents: {
+            ...bookingContentsAsStrings,
+            headerMessage: `A ${config.displayName} service approval is required for this request.`,
+          },
+          targetEmail: recipient,
+          status: BookingStatusLabel.REQUESTED,
+          subjectStatusOverride: config.subjectStatus,
+          eventTitle: bookingContents.title || "",
+          requestNumber: bookingContents.requestNumber,
+          body: "",
+          replyTo: requesterEmail,
+          tenant,
+          schemaName: emailConfig.schemaName,
+        }),
+      );
+    },
+  );
+
+  await Promise.all(emailJobs);
+};
 
 async function handleBookingApprovalEmails(
   isAutoApproval: boolean,
@@ -908,6 +1006,29 @@ export async function POST(request: NextRequest) {
           },
         );
         // Don't fail the entire booking if XState save fails
+      }
+    }
+
+    if (isMediaCommons) {
+      const servicesRequested = getMediaCommonsServices(data);
+      try {
+        await notifyServiceApproversForRequestedServices(
+          calendarEventId,
+          email,
+          servicesRequested,
+          tenant,
+        );
+      } catch (notificationError) {
+        console.error(
+          `🚨 SERVICE APPROVER NOTIFICATION FAILED [${tenant?.toUpperCase()}]:`,
+          {
+            calendarEventId,
+            error:
+              notificationError instanceof Error
+                ? notificationError.message
+                : String(notificationError),
+          },
+        );
       }
     }
